refactor: blocks
