+++ example/package.json
@@ -9,7 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@udus/notion-renderer": "file:..",
+    "@udus/notion-renderer": "^1.0.0-alpha.52",
     "next": "14.0.4",
     "react": "^18",
     "react-dom": "^18"
