+++ example/src/components/notion-renderer.ts
@@ -1,0 +1,2 @@
+"use client";
+export * from "@udus/notion-renderer/components";
