+++ src/components/Page/Page.stories.tsx
@@ -1,13 +1,6 @@
-import BlocksObject from "../Blocks/BlockList.json";
-import { OpenedHeading1 } from "../Blocks/Custom/OpenedHeading1.js";
-import { OpenedHeading2 } from "../Blocks/Custom/OpenedHeading2.js";
-import { OpenedHeading3 } from "../Blocks/Custom/OpenedHeading3.js";
-import { OpenedToggle } from "../Blocks/Custom/OpenedToggle.js";
-
 import { Page } from "./Page.js";
 import PageMetaObjectData from "./PageMeta.json";
 
-import type { BlockBlockObject } from "../../types/notion/block/block.js";
 import type { PageObject } from "../../types/notion/page.js";
 import type { Meta, StoryObj } from "@storybook/react";
 
@@ -24,7 +17,6 @@ type Story = StoryObj<typeof Page>;
 export const Default: Story = {
   args: {
     page: PageMetaObjectData as PageObject,
-    blocks: BlocksObject as Array<BlockBlockObject>,
     displayProperties: [
       "Status",
       "Multi-select",
@@ -60,13 +52,6 @@ export const Default: Story = {
 export const Custom: Story = {
   args: {
     page: PageMetaObjectData as PageObject,
-    blocks: BlocksObject as Array<BlockBlockObject>,
-    blockMapper: {
-      toggle: OpenedToggle,
-      heading_1: OpenedHeading1,
-      heading_2: OpenedHeading2,
-      heading_3: OpenedHeading3,
-    },
   },
   parameters: {
     controls: {
